Refactored ApplicationInformation
